requirements.txt
Flask==2.3.3
insightface==0.7.3
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
onnxruntime==1.16.0
gunicorn==21.2.0